requirements.txt
alembic==1.9.4
anyio==3.7.1
bidict==0.22.1
cachelib==0.10.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
exceptiongroup==1.1.2
Flask==2.2.3
flask-http-middleware==0.2.1
Flask-Migrate==4.0.4
Flask-Pydantic==0.11.0
Flask-Script==2.0.6
Flask-Session==0.4.0
flask-sock==0.6.0
Flask-SocketIO==5.3.2
Flask-SQLAlchemy==3.0.3
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.7.0
Mako==1.2.4
MarkupSafe==2.1.2
Pillow==10.0.0
psycopg2-binary==2.9.5
pydantic==1.10.5
python-dotenv==0.21.1
python-engineio==4.3.4
python-socketio==5.7.2
requests==2.28.2
simple-websocket==0.10.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.30.0
typing_extensions==4.5.0
urllib3==1.26.14
validation==0.8.3
Werkzeug==2.2.3
wsproto==1.2.0
zipp==3.12.0
